fix(uninstall.sh): remove CELLAR_DIR only for macOS

+++ scripts/uninstall.sh
@@ -247,9 +247,11 @@ if [ -d "$USR_LIB_DIR" ]; then
     maybe_sudo rm -rf "$USR_LIB_DIR" || warn_message "Failed to remove Suricata lib folder."
 fi
 
-if [ -d "$CELLAR_DIR" ]; then
-    info_message "Removing Suricata cellar folder..."
-    maybe_sudo rm -rf "$CELLAR_DIR" || warn_message "Failed to remove Suricata cellar folder."
+if [ "$(uname -s)" = "Darwin" ]; then
+    if [ -d "$CELLAR_DIR" ]; then
+        info_message "Removing Suricata cellar folder..."
+        maybe_sudo rm -rf "$CELLAR_DIR" || warn_message "Failed to remove Suricata cellar folder."
+    fi
 fi
 
 # Only run on Linux
